requirements.txt
alpaca-py>=0.28,<1.0
pandas>=2.0,<3.0
pyzmq>=25,<27
msgpack>=1.0,<2

